Add: texture sturcture separation
